corrected splitted api and web views
