+++ lib/lovelace_web/views/api_success_view.ex
@@ -1,4 +1,4 @@
-defmodule LovelaceWeb.SuccessView do
+defmodule LovelaceWeb.APISuccessView do
   use LovelaceWeb, :view
 
   def render("success.json", %{body: body}) do
