+++ lib/lovelace_web/views/error_view.ex
@@ -3,26 +3,14 @@ defmodule LovelaceWeb.ErrorView do
 
   # If you want to customize a particular status code
   # for a certain format, you may uncomment below.
-  # def render("500.json", _assigns) do
-  #   %{errors: %{detail: "Internal Server Error"}}
+  # def render("500.html", _assigns) do
+  #   "Internal Server Error"
   # end
 
   # By default, Phoenix returns the status message from
-  # the template name. For example, "404.json" becomes
+  # the template name. For example, "404.html" becomes
   # "Not Found".
   def template_not_found(template, _assigns) do
-    %{errors: %{detail: Phoenix.Controller.status_message_from_template(template)}}
-  end
-
-  def render("generic_error.json", %{body: body}) do
-    body
-  end
-
-  def render("404.json", _) do
-    %{message: "not_found"}
-  end
-
-  def render("500.json", _assigns) do
-    %{message: "internal_server_error"}
+    Phoenix.Controller.status_message_from_template(template)
   end
 end
